backend/requirements.txt
# backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.1.0
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
pytesseract==0.3.10
pdf2image==1.16.3
python-docx==1.1.0